requirement.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
cryptography==46.0.3
deprecation==2.1.0
dj-database-url==3.0.1
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
multidict==6.7.0
packaging==25.0
pi==0.1.2
postgrest==2.24.0
propcache==0.4.1
psycopg2==2.9.11
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
realtime==2.24.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==2.24.0
StrEnum==0.4.15
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
websockets==15.0.1
yarl==1.22.0
